Remove inkscape dependency, just use SVGs
The inkscape dependency was used to convert some images (github/twitter/website
logos for speakers and sponsor logos) from SVG to PNG.  However, it proves to
be an annoying dependency on Mac OS X.

This patch changes just makes the website use the SVGs directly.  It seems to
work fine in Firefox and Chromium.  I had to set an explicit width/height for
the HSR logo only.